[Fix] #48 #50 쇼츠, 씬스틸러 영상 재생 시각 자료형 수정

+++ Seetube/Seetube/Data/Network/DTO/Youtuber/SceneStealerResultDTO.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct SceneStealerSceneDTO: Decodable, DomainConvertible {
     let thumbnailURL: String
-    let startTime: Int
-    let endTime: Int
+    let startTime: Float
+    let endTime: Float
     let percentageOfConcentration: Int
     
     func toDomain() -> SceneStealerScene {
